% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-functions.R
\name{add_trans_ci}
\alias{add_trans_ci}
\title{Add transition probabilities confidence intervals}
\usage{
add_trans_ci(
  newdata,
  object,
  nsim = 100L,
  alpha = 0.05,
  time_var = NULL,
  interval_length = "intlen",
  transition = "transition",
  ...
)
}
\description{
Add transition probabilities confidence intervals
}
\keyword{internal}
