% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform-fda.R
\name{as_fped}
\alias{as_fped}
\title{Transform data with one row per subject to functional PED format}
\usage{
as_fped(formula, data, data_orig, cut = NULL, ..., max.end = FALSE,
  ll_fun = function(te, t) {     te <= t }, keep = c("id", "tstart",
  "eta_base", "eta_wce", "eta"))
}
\arguments{
\item{formula}{a model formula}

\item{data}{a data frame}

\item{cut}{the vector of timepoints to cut at}

\item{...}{Further arguments passed to \code{\link[survival]{survSplit}}}

\item{max.end}{logical. Should the last interval span until the last 
observed censoring or event time (if larger than the largest specified 
cut point).}

\item{ll_fun}{Function with two arguments that indicates how lag-lead matrix 
should be contructed. First argument is the time on the scale of the exposure, 
second argument is the time on the scale of the follow-up. Internally the 
lag lead matrix will be constructed by calling \code{t(outer(te_vec, t_vec, ll_fun))}.}
}
\value{
A data frame class \code{ped} in piece-wise exponential data format.
}
\description{
Transform data with one row per subject to functional PED format
}
\seealso{
\code{\link[survival]{survSplit}}
}
\keyword{internal}
