% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nest-utils.R
\name{as_matdf}
\alias{as_matdf}
\title{Transform a nested data frame to data frame with matrix columns}
\usage{
as_matdf(nested_df)
}
\arguments{
\item{nested_df}{A nested data frame with list columns that contain 
a matrix per row.}
}
\description{
Transform a nested data frame to data frame with matrix columns
}
