% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-ped.R
\name{as_ped.nested_fdf}
\alias{as_ped.nested_fdf}
\title{Transform data to Piece-wise Exponential Data (PED)}
\usage{
\method{as_ped}{nested_fdf}(data, formula, ...)
}
\arguments{
\item{data}{a data frame}

\item{formula}{a model formula}

\item{...}{Further arguments passed to \code{\link[survival]{survSplit}}}
}
\value{
A data frame class \code{ped} in piece-wise exponential data format.
}
\description{
Transform data to Piece-wise Exponential Data (PED)
}
\seealso{
\code{\link[survival]{survSplit}}
}
