% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-ped-cr.R
\name{as_ped_cr}
\alias{as_ped_cr}
\title{Transform competing risks data to Piece-wise Exponential Data (PED)}
\usage{
as_ped_cr(
  data,
  formula,
  cut = NULL,
  max_time,
  censor_code = 0L,
  combine = TRUE,
  ...
)
}
\arguments{
\item{data}{Either an object inheriting from data frame or in case of
time-dependent covariates a list of data frames (of length 2), where the first data frame
contains the time-to-event information and static covariates while the second
(and potentially further data frames) contain information on time-dependent
covariates and the times at which they have been observed.}

\item{formula}{A two sided formula with a \code{\link[survival]{Surv}} object
on the left-hand-side and covariate specification on the right-hand-side (RHS).
The RHS can be an extended formula, which specifies how TDCs should be transformed
using specials \code{concurrent} and \code{cumulative}.}

\item{cut}{Either a numeric vector or a list of numeric vectors.
If provided as list, the length of the list must equal the number of
competing risks.
Specifies split points, used to partition the follow up into intervals.
If unspecified, all unique event times will be used.
This argument interacts with argument \code{combine}. If \code{cut} is unspecified
and \code{combine} is \code{FALSE}, each risk has a different set of cuts.
If set to \code{combine = TRUE}, the cut points are a union of unique event
times across all competing events.}

\item{max_time}{If \code{cut} is unspecified, this will be the last
possible event time. All event times after \code{max_time}
will be administratively censored at \code{max_time}.}

\item{censor_code}{Either a string or integer (depending on data) with out
outlines which level in the status variable is associated with censorship.}

\item{combine}{A logical value. Defaults to \code{TRUE}. Indicates which
output type should be used. \code{TRUE} results in a joint PED object for
all risks, i.e., PED objects for each cause are stacked and returned as one data frame.
\code{FALSE} results in a named list of PED objects.}

\item{...}{Further arguments passed to the \code{as_ped} function or its
methods (\code{data.frame} method) and eventually to
\code{\link[survival]{survSplit}}}
}
\value{
A data frame class \code{ped} in piece-wise exponential data format
with an additional column indicating which cause is observed (if output ==
"data.frame".) or a named list of single risk data.frames (if output ==
"list).
}
\description{
Transformation of competing risks data to Piece-wise Exponantial Data (PED).
The data transformation works equivantly to the single event case (see
\code{\link{as_ped}} for details). More details about data transformation are
also provided in the \href{https://adibender.github.io/pammtools//articles/data-transformation.html}{data-transformation
vignette}.
}
\details{
This function wraps the \code{as_ped} function and repeats the data
transfomation for each single risk. The data transformation procedure can be
different for all associated risks, e.g. the user can supply distinct cut
points for each single risk. This function returns either a data.frame
which complies with the data.frames proposed
\href{https://arxiv.org/abs/2006.15442}{here} or a list of single risk
PED objects. (This depends on the argument "output".)
Currently our function only supports cause-specific competing
risks. That means that for each single risk a competing event is treated
(and recoded) as censoring event.
}
\examples{
data("pbc", package = "survival")
pbc_tmp <- pbc[c(1, 2, 5), ]
# creates cause specific data sets with cause specific, custom interval split points
# and stacks them
pbc_tmp \%>\% as_ped_cr(Surv(time, status) ~ age + sex,
                      cut = list(c(0, 200, 500), c(0, 500, 1550)))
# returns cause specific data sets as a list
pbc_tmp \%>\% as_ped_cr(Surv(time, status) ~ age + sex, combine = FALSE)
# provide censoring code if status is not numeric
pbc_tmp$status <- c("death", "discharge", "censoring")
# by default, a union of event times for all causes will be used as split points
pbc_tmp \%>\% as_ped_cr(Surv(time, status) ~ age + sex, censor_code = "censoring")
pbc_tmp \%>\% as_ped_cr(Surv(time, status) ~ age + sex, combine = FALSE,
                      censor_code = "censoring")
}
\author{
Philipp Kopper
}
