% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidiers.R
\name{calc_ci}
\alias{calc_ci}
\title{Calculate confidence intervals}
\usage{
calc_ci(ftab)
}
\arguments{
\item{ftab}{A table with two columns, containing coefficients in the first 
column and standard-errors in the second column.}
}
\description{
Given 2 column matrix or data frame, returns 3 column data.frame 
with coefficient estimate plus lower and upper borders of the 
95% confidence intervals.
}
