% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgcv-utils.R
\name{compute_cum_diff}
\alias{compute_cum_diff}
\title{Calculate difference in cumulative hazards and respective standard errors}
\usage{
compute_cum_diff(d1, d2, model)
}
\arguments{
\item{d1}{A data set used as \code{newdata} in \code{predict.gam}}

\item{d2}{See \code{d1}}

\item{model}{A model object for which a predict method is implemented which
returns the design matrix (e.g., \code{mgcv::gam}).}
}
\description{
Uses Delta method to calculate standard errors for cumulative hazard
differences based on coefficients and standard errors of log hazard
(code adapted from fabian-s used in consulting project).
}
\keyword{internal}
