% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumulative-coefficient.R
\name{compute_cumu_diff}
\alias{compute_cumu_diff}
\title{Calculate difference in cumulative hazards and respective standard errors}
\usage{
compute_cumu_diff(d1, d2, model, alpha = 0.05, nsim = 100L)
}
\arguments{
\item{d1}{A data set used as \code{newdata} in \code{predict.gam}}

\item{d2}{See \code{d1}}

\item{model}{A model object for which a predict method is implemented which
returns the design matrix (e.g., \code{mgcv::gam}).}
}
\description{
CIs are calculated by sampling coefficients from their posterior and
calculating the cumulative hazard difference \code{nsim} times. The CI
are obtained by the 2.5\\% and 97.5\\% quantiles.
}
\keyword{internal}
