% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{construct_newdata}
\alias{construct_newdata}
\title{Construct a data frame suitable for prediction}
\usage{
construct_newdata(ped, expand = NULL)
}
\arguments{
\item{ped}{An object of class \code{ped} as returned by \code{\link[pam]{split_data}}.}

\item{expand}{A charachter vector of column names in \code{ped}.}
}
\description{
Given a data set of class \code{ped}, returns a data frame that can be used 
as \code{newdata} argument in a call to \code{predict} and similar functions. 
When \code{expand==NULL} falls back to \code{\link{ped_info}}.
}
