% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{daily}
\alias{daily}
\title{Time-dependent covariates of the \code{\link{patient}} data set.}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 18797 rows and 4 columns.
}
\usage{
daily
}
\description{
This data set contains the time-dependent covariates (TDCs) for the \code{\link{patient}}
data set. Note that nutrition was protocoled for at most 12 days after
ICU admission. The data set includes:
\describe{
\item{CombinedID}{Unique patient identifier. Can be used to merge with
\code{\link{patient}} data}
\item{Study_Day}{The calendar (!) day at which calories (or proteins) were
administered}
\item{caloriesPercentage}{The percentage of target calories supplied to the
patient by the ICU staff}
\item{proteinGproKG}{The amount of protein supplied to the patient by the
ICU staff}}
}
\keyword{datasets}
