% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_fixed}
\alias{extract_fixed}
\alias{extract_fixed.gam}
\alias{extract_fixed.coxph}
\title{Extract fixed coefficient table from model object}
\usage{
extract_fixed(x, ...)

\method{extract_fixed}{gam}(x, ...)

\method{extract_fixed}{coxph}(x, ...)
}
\arguments{
\item{x}{A model object.}

\item{...}{Currently not used.}
}
\description{
Given a model object, returns data frame with columns \code{variable},
\code{coef} (coefficient), \code{lower} (lower 95\% CI) and 
\code{upper} (upper 95% CI).
}
