% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform-fda.R
\name{extract_one}
\alias{extract_one}
\title{Exctract id with most observations from fped object}
\usage{
extract_one(fped)
}
\arguments{
\item{fped}{A functional PED.}
}
\description{
Exctract id with most observations from fped object
}
\keyword{internal}
