% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-pexp.R
\name{fcumu}
\alias{fcumu}
\title{A formula special used to handle cumulative effect specifications}
\usage{
fcumu(..., by = NULL, f_xyz, ll_fun)
}
\description{
Can be used in the second part of the formula specification provided
to \code{\link[pammtools]{sim_pexp}} and should only be used in this
context.
}
\keyword{internal}
