% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula-utils.R
\name{get_lhs_vars}
\alias{get_lhs_vars}
\alias{get_rhs_vars}
\title{Extract variables from the left-hand-side of a formula}
\usage{
get_lhs_vars(formula)

get_rhs_vars(formula)
}
\arguments{
\item{formula}{A \code{\link{formula}} object.}
}
\description{
Extract variables from the left-hand-side of a formula

Extract variables from the right-hand side of a formula
}
