% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula-utils.R
\name{func}
\alias{func}
\title{A formula special for defining functional covariates}
\usage{
func(..., te_var, ll_fun = function(t, te) {     t >= te }, suffix = NULL)
}
\description{
A formula special for defining functional covariates
}
\keyword{internal}
