% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-helpers.R
\name{get_coefs}
\alias{get_coefs}
\alias{get_coefs.default}
\alias{get_coefs.scam}
\title{Extract model coefficients on the scale of the design matrix}
\usage{
get_coefs(object, ...)

\method{get_coefs}{default}(object, ...)

\method{get_coefs}{scam}(object, ...)
}
\arguments{
\item{object}{A fitted model object.}

\item{...}{Further arguments passed to methods.}
}
\description{
Returns the coefficient vector \code{coefs} such that
\code{make_X(object, newdata) \%*\% coefs} yields the linear predictor.
For most models this is simply \code{coef(object)}. For \code{scam} models,
however, \code{coef()} returns the coefficients on the underlying
unconstrained scale, while the linear predictor is calculated from the
re-parametrized (partially exponentiated) coefficients
\code{object$coefficients.t}.
}
\keyword{internal}
