% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hazard.R
\name{get_cumhazard}
\alias{get_cumhazard}
\title{Calculate cumulative hazard}
\usage{
get_cumhazard(newdata, object, ci = TRUE, time_variable = NULL,
  interval_length = quo(intlen), ...)
}
\arguments{
\item{newdata}{ A data frame or list containing the values of the model covariates at which predictions
  are required. If this is not provided then predictions corresponding to the
                original data are returned. If \code{newdata} is provided then
                it should contain all the variables needed for prediction: a
                warning is generated if not. See details for use with \code{link{linear.functional.terms}}. }

\item{object}{ a fitted \code{gam} object as produced by \code{gam()}.
                }

\item{ci}{Logical indicating whether to iclude confidence intervals. Defaults
to \code{TRUE}}

\item{time_variable}{Name of the variable used for the baseline hazard. If
not given, defaults to \code{"tend"} for \code{\link[mgcv]{gam}} fits, else
\code{"interval"}. The latter is assumed to be a factor, the former
numeric.}

\item{interval_length}{\code{quosure} providing the name of the variable in
newdata containing the interval lengths. Defaults to \code{intlen}.}

\item{...}{Further arguments passed to \code{\link[mgcv]{predict.gam}} and
\code{\link{get_hazard}}}
}
\description{
Calculate cumulative hazard
}
