% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumulative-coefficient.R
\name{get_cumu_coef.gam}
\alias{get_cumu_coef.gam}
\title{Cumulative coefficients}
\usage{
\method{get_cumu_coef}{gam}(model, data, terms, ...)
}
\arguments{
\item{model}{Object from which to extract cumulative coefficients.}

\item{data}{Additional data if necessary.}

\item{terms}{A character vector of variables for which the cumulative
coefficient should be calculated.}

\item{...}{Further arguments passed to \code{\link[pammtools]{add_cumu_hazard}}.}
}
\description{
These functions are designed to mimick the cumulative coefficients
usually used in additive hazards models (Aalen model) to depict (time-varying)
covariate effects. For PAMMs, these can be interpreted as the differences
between the cumulative hazard rates where all covariates except one have the
same values. For the a numeric covariate of interest we calculate
Lambda(t|x+1) - Lambda(t|x).  For non-numeric covariates of interest
the cumulative hazard of the reference level is substracted from
the cumulative hazards eveluated at all non reference levels. Standard
errors are calculated using the delta method.
}
