% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumulative-effect.R
\name{get_cumu_eff}
\alias{get_cumu_eff}
\title{Calculate cumulative effect for all time-points}
\usage{
get_cumu_eff(data, model, term, z1, z2 = NULL, se_mult = 2)
}
\arguments{
\item{model}{A suitable model object which will be used to estimate the
partial effect of \code{term}.}

\item{z1}{The expoure profile for which to calculate the cumulative effect.
Can be either a singel number or a vector of same length as unique observation
time points.}

\item{z2}{If provided, calculated cumulative effect is for the difference
between the two exposure profiles (g(z1,t)-g(z2,t)).}
}
\description{
Calculate cumulative effect for all time-points
}
\keyword{internal}
