% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-cut-points.R
\name{get_cut}
\alias{get_cut}
\alias{get_cut.default}
\title{Obtain interval break points}
\usage{
get_cut(data, formula, cut = NULL, ...)

\method{get_cut}{default}(data, formula, cut = NULL, max_time = NULL,
  ...)
}
\description{
Obtain interval break points
}
\keyword{internal}
