% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trafo-cumulative.R
\name{get_func}
\alias{get_func}
\alias{expand_func}
\title{Expand time-dependent covariates to functionals}
\usage{
get_func(data, formula)

expand_func(data, func)
}
\arguments{
\item{data}{Data frame (or similar) in which variables specified in ...
will be looked for}

\item{func}{Single \code{\link{func}} term \code{\link{get_func}} \code{formula}
evaluated.}

\item{...}{One sided formulas of the form ~func(t,te,x) that specify the
type of cumulative effect desired (see examples). Possible specifications are
\itemize{
\item \code{func(t,te,x)}: The most general specification.
\item \code{func(t, t-te,x)}: A time-varying DLNM.
\item \code{func(t-te,x}: A DLNM.
\item \code{func(t-te,by=x) A WCE}}}
}
\description{
Given formula specification on how time-dependent covariates affect the
outcome, creates respective functional covariate as well as auxilliary
matrices for time/lattency etc.
}
\keyword{internal}
