% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group-helpers.R
\name{get_grpvars}
\alias{get_grpvars}
\title{Extract character vector of grouping variables}
\usage{
get_grpvars(data)
}
\arguments{
\item{data}{A data frame (potentially `grouped_df`) from which to extract 
names of grouping variables.}
}
\value{
A character vector (of length 0 if no grouping variables present).
}
\description{
Extract character vector of grouping variables
}
