% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervalInformation.R
\name{get_intervals}
\alias{get_intervals}
\title{Given breaks, return intervals in which times vector falls}
\usage{
get_intervals(x, times, ...)
}
\arguments{
\item{x}{An object from which interval information can be obtained, 
see \code{\link{int_info}}.}

\item{times}{A vector of times for which corresponding interval information 
should be returned.}

\item{...}{Further arguments passed to \code{\link[base]{findInterval}}.}
}
\value{
A \code{data.frame} containing information on intervals in which
values of x fall.
}
\description{
Given breaks, return intervals in which times vector falls
}
\examples{
set.seed(111018)
brks <- c(0, 4.5, 5, 10, 30)
int_info(brks)
x <- runif(3, 0, 30)
get_intervals(brks, x, left.open=TRUE)

}
\seealso{
findInterval int_info
}
