% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lag-lead-utils.R
\name{get_laglead}
\alias{get_laglead}
\alias{get_laglead.default}
\alias{get_laglead.data.frame}
\title{Construct data that represents a lag-lead window}
\usage{
get_laglead(x, ...)

\method{get_laglead}{default}(x, te, ll_fun, ...)

\method{get_laglead}{data.frame}(x, ...)
}
\arguments{
\item{x}{Either a numeric vector of follow-up cut points or a suitable object.}

\item{...}{Further arguments passed to methods.}

\item{te}{A vector of exposure times}
}
\description{
Constructs lag-lead window data set from raw inputs or from data objects
with suitable information stored in attributes.
}
