% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terms-info.R
\name{get_plotinfo}
\alias{get_plotinfo}
\title{Extract plot information for all special model terms}
\usage{
get_plotinfo(x, ...)
}
\arguments{
\item{x}{ a fitted \code{gam} object as produced by \code{gam()}.}

\item{...}{Further arguments passed to \code{\link[mgcv]{plot.gam}}}
}
\description{
Given a \code{mgcv} \code{\link[mgcv]{gamObject}}, returns the information
used for the default plots produced by \code{\link[mgcv]{plot.gam}}.
}
