% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-functions.R
\name{get_sim_cumu}
\alias{get_sim_cumu}
\title{helper function for add_trans_ci}
\usage{
get_sim_cumu(newdata, interval_length = "intlen", ...)
}
\description{
helper function for add_trans_ci
}
\keyword{internal}
