% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula-utils.R
\name{get_tdc_form}
\alias{get_tdc_form}
\title{Extract variables from the left-hand-side of a formula}
\usage{
get_tdc_form(
  formula,
  data = NULL,
  tdc_specials = c("concurrent", "cumulative"),
  invert = FALSE
)
}
\arguments{
\item{formula}{A \code{\link{formula}} object.}
}
\description{
Extract variables from the left-hand-side of a formula

Extract variables from the right-hand side of a formula
}
\keyword{internal}
