% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula-utils.R
\name{get_tdc_vars}
\alias{get_tdc_vars}
\title{Extract variables from the left-hand-side of a formula}
\usage{
get_tdc_vars(formula, specials = "cumulative", data = NULL)
}
\arguments{
\item{formula}{A \code{\link{formula}} object.}
}
\description{
Extract variables from the left-hand-side of a formula

Extract variables from the right-hand side of a formula
}
\keyword{internal}
