% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-helpers.R
\name{get_Vp}
\alias{get_Vp}
\alias{get_Vp.default}
\alias{get_Vp.gam}
\alias{get_Vp.scam}
\title{Extract the (Bayesian) covariance matrix of the model coefficients}
\usage{
get_Vp(object, ...)

\method{get_Vp}{default}(object, ...)

\method{get_Vp}{gam}(object, ...)

\method{get_Vp}{scam}(object, ...)
}
\arguments{
\item{object}{A fitted model object.}

\item{...}{Further arguments passed to methods.}
}
\description{
Returns the covariance matrix that matches the coefficients returned by
\code{\link{get_coefs}}. For \code{mgcv} models this is the Bayesian
posterior covariance matrix \code{object$Vp}, for \code{scam} models the
covariance matrix of the re-parametrized coefficients \code{object$Vp.t}
and \code{vcov(object)} otherwise.
}
\keyword{internal}
