% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz-elra.R
\name{gg_partial}
\alias{gg_partial}
\alias{gg_slice}
\title{Visualize effect estimates for specific covariate combinations}
\usage{
gg_partial(data, mod, term, ...)

gg_slice(data, mod, term, ...)
}
\description{
Depending on the plot type and input, creates either a bivariate effect
surface or slices through 1 dimension.
}
\keyword{internal}
\keyword{internaldo}
