% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience-plots.R
\name{gg_re}
\alias{gg_re}
\title{Plot Normal QQ plots for random effects}
\usage{
gg_re(x, ...)
}
\arguments{
\item{x}{A model object.}

\item{...}{Currently not used.}
}
\description{
Plot Normal QQ plots for random effects
}
\examples{
data("lung", package="survival")
lung$inst <- as.factor(lung$inst) # for mgcv
ped <- split_data(Surv(time, status)~ph.ecog + inst, data=lung, id="id")
pam <- mgcv::gam(ped_status ~ s(tend) + ph.ecog + s(inst, bs="re"),
	data=ped, family=poisson(), offset=offset)
gg_re(pam)
}
\seealso{
\code{\link{tidy_re}}
}
