% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience-plots.R
\name{gg_smooth.default}
\alias{gg_smooth.default}
\title{Plot smooth 1d terms of gam objects}
\usage{
\method{gg_smooth}{default}(x, fit, ...)
}
\arguments{
\item{x}{A data frame or object of class \code{pamm}.}

\item{fit}{A fitted model object.}

\item{...}{Further arguments passed to \code{\link{get_terms}}}
}
\value{
A \code{\link[ggplot2]{ggplot2}} object.
}
\description{
Given a gam model this convenience function returns a plot of all 
smooth terms contained in the model. If more than one smooth is present, the 
different smooth are faceted.
}
