% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conveniencePlots.R
\name{gg_tensor}
\alias{gg_tensor}
\title{Plot tensor product effects}
\usage{
gg_tensor(fit, ci = FALSE, ...)
}
\arguments{
\item{fit}{A fitted object of class \code{gam}.}

\item{...}{Currently ignored.}
}
\description{
Given a gam model this convenience function returns a \code{ggplot2} object 
depicting 2d smooth terms specified in the model as heat/contour plots. If 
more than one 2d smooth term is present individual terms are faceted.
}
\examples{
library(mgcv)
g <- gam(Sepal.Length ~ te(Sepal.Width, Petal.Length), data=iris)
gg_tensor(g)
}
\seealso{
\code{\link{tidy_smooth2d}}
}
