% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula-specials.R
\name{has_special}
\alias{has_special}
\title{Expand time-dependent covariates to functionals}
\usage{
has_special(formula, special = "func")
}
\description{
Given formula specification on how time-dependent covariates affect the
outcome, creates respective functional covariate as well as auxiliary
matrices for time/latency etc.
}
