% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lag-lead-helpers.R
\name{Lmat}
\alias{Lmat}
\title{Extend instance of Lag-Lead matrix to whole data set}
\usage{
Lmat(LL, id)
}
\arguments{
\item{LL}{An instance of the Lag-Lead matrix (for all time points of the follow-up)}

\item{id}{The id vector of the data set in PED format.}
}
\description{
Extend instance of Lag-Lead matrix to whole data set
}
