% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lag-lead-helpers.R
\name{Lsimp}
\alias{Lsimp}
\title{Create one instance of the Lag-Lead matrix}
\usage{
Lsimp(te_vec, t_vec, ll_fun = function(te, t) {     te <= t })
}
\arguments{
\item{te_vec}{The vector of exposure times (times at which Change in TDC occured)}

\item{t_vec}{The vector of event/split-times on the scale of the follow-up.}

\item{ll_fun}{Function with two arguments that indicates how lag-lead matrix 
should be contructed. First argument is the time on the scale of the exposure, 
second argument is the time on the scale of the follow-up. Internally the 
lag lead matrix will be constructed by calling \code{t(outer(te_vec, t_vec, ll_fun))}.}
}
\description{
Create one instance of the Lag-Lead matrix
}
