% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct-newdata.R
\name{make_newdata.ped}
\alias{make_newdata.ped}
\title{Construct a data frame suitable for prediction}
\usage{
\method{make_newdata}{ped}(x, ..., expand = NULL, n = 50L)
}
\arguments{
\item{x}{A data frame (or object that inherits from \code{data.frame}).}

\item{...}{Further specifications of variables that should be set
to a specific values.}

\item{expand}{A character vector of column names in \code{ped}.}

\item{n}{If \code{expand} specified, respective variables will be expanded 
in \code{n} values from minimum to maximum.}
}
\description{
Given a data set, returns a data frame type object that can be used
as \code{newdata} argument in a call to \code{predict} and similar functions.
}
\details{
Extracts information from \code{ped}, using
  \code{\link{sample_info}}. If variables are specified with specific values
  in \code{...}, the values in from \code{sample_info} will be overwritten.
  If variables are provided in \code{expand}, these will be expanded from \code{min} to
  \code{max} using in \code{n} equidistant steps.
}
