% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct-newdata.R
\name{make_newdata}
\alias{make_newdata}
\title{Construct a data frame suitable for prediction}
\usage{
make_newdata(x, ..., expand = NULL, length.out = 50L)
}
\arguments{
\item{x}{A data frame (or object that inherits from \code{data.frame}).}

\item{...}{Further specifications of variables that should be set
to a specific value.}

\item{expand}{A character vector of column names in \code{ped}.}

\item{length.out}{desired length of the sequence.  A
    non-negative number, which for \code{seq} and \code{seq.int} will be
    rounded up if fractional.}
}
\description{
Given a data set, returns a data frame type object that can be used
as \code{newdata} argument in a call to \code{predict} and similar functions.
}
\details{
Extracts information from \code{ped}, using
  \code{\link{sample_info}}. If variables are specified with specific values
  in \code{...}, the values in from \code{sample_info} will be overwritten.
  If variables are provided in \code{expand}, these will be expanded from \code{min} to
  \code{max} using in \code{length.out} equidistant steps.
}
\examples{
\dontrun{
library(dplyr)
iris \%>\% make_newdata()
iris \%>\% make_newdata(Sepal.Length=5)
iris \%>\% make_newdata(Sepal.Length=c(5, 10), Sepal.Width=c(5, 5.1, 5.2))
iris \%>\% make_newdata(Sepal.Length=c(5, 10), expand="Sepal.Width", length.out=5)
iris \%>\% group_by(Species) \%>\%
  make_newdata(Sepal.Length=c(5, 10), expand="Sepal.Width", length.out=5) \%>\%
  print(n=30)
}
}
