% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-ped-cr.R
\name{make_numeric}
\alias{make_numeric}
\title{Conversion to numeric risks}
\usage{
make_numeric(data, status_str, censor_code)
}
\arguments{
\item{data}{A data.frame of class ped_cr that features time-to-event data
and convariates. (see https://adibender.github.io/pammtools/)}

\item{status_str}{A character indicating the colname of the status variable.}

\item{censor_code}{A value (character/numeric/factor) indicating as what
censoring has been coded.}
}
\value{
data with (converted) numeric status column
}
\description{
This function converts character or factor status into numeric ones.
This is necessary for further processing.
}
