% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-functions.R
\name{make_X.gam}
\alias{make_X.gam}
\title{Create model matrix from model object}
\usage{
\method{make_X}{gam}(object, newdata, ...)
}
\arguments{
\item{object}{a model object}
}
\description{
Create model matrix from model object
}
\keyword{internal}
