% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation-functions.R
\name{make_X}
\alias{make_X}
\title{Construct full data set with time-dependent covariate (TDC)}
\usage{
make_X(n = 500L, m = 30L, h0 = -3, f0 = function(t) (0 * t),
  fte = function(te) (-2 * te + 2.5 * te^2))
}
\arguments{
\item{n}{Number of subjects to simulate.}

\item{m}{= the number of time-points at which TDC is observed.}

\item{h0}{A constant component of the baseline hazard.}

\item{f0}{A function of t, representing the smooth part of the baseline hazard.}

\item{fte}{A function of the exposure time, representing the functional 
effect of the timing of the TDC.}
}
\description{
Construct full data set with time-dependent covariate (TDC)
}
\keyword{internal}
