% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-functions.R
\name{make_X.scam}
\alias{make_X.scam}
\title{Create design matrix from a suitable object}
\usage{
\method{make_X}{scam}(object, newdata, ...)
}
\arguments{
\item{object}{A suitable object from which a design matrix can be generated.
Often a model object.}

\item{newdata}{A data frame from which design matrix will be constructed}
}
\description{
Create design matrix from a suitable object
}
\keyword{internal}
