% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{mk_ndf}
\alias{mk_ndf}
\title{Create new data for cumulative effects models}
\usage{
mk_ndf(x, ...)
}
\arguments{
\item{x}{An object of class \code{fped}.}

\item{...}{Specification of variables and their values. Functions can be used,
e.g., \code{x1 = seq_range(x1)}.}
}
\description{
Experimental new implementation of \code{make_newdata} for \code{fped}
objects.
}
\keyword{internal}
