% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pec-helpers.R
\name{predictSurvProb.pam}
\alias{predictSurvProb.pam}
\alias{pec_cv}
\alias{pec_pam}
\alias{tidy_pec}
\title{Predict survival probabilities at specified time points}
\usage{
predictSurvProb.pam(object, newdata, times, ...)

pec_cv(data, formula_pam, formula_ped = Surv(time, status) ~ .,
  method = pam, k = 10L, ...)

pec_pam(data, idx_train, idx_test, formula_ped, formula_pam, times = NULL,
  n = 50L, ...)

tidy_pec(pec)
}
\arguments{
\item{object}{A PAM model object.}

\item{newdata}{Data containing covariate information only, for which to 
predict survival probabilities.}

\item{times}{Time points for which survival probabilities should be predicted.}

\item{...}{Arguments passed to \code{\link[pam]{int_info}}}

\item{data}{The data in standard time-to-event data format which can be tansformed 
in PED (see \code{\link{split_data}}).}

\item{formula_pam}{The formula used to fit the PAM (see \code{\link{pam}}).}

\item{formula_ped}{The formula used to transform data to PED (see 
\code{\link{split_data}}).}

\item{method}{The function to which \code{formula_pam} will be passed to fit the 
model in the k-th fold.}

\item{k}{The number of folds to be used in cross-validation.}

\item{idx_train}{Index of rows that indicate training set.}

\item{idx_test}{Index of rows that indicate test set.}

\item{n}{If \code{times} argument not specified, will create a sequence 
from minimal to maximal time of this length.}

\item{pec}{An object of class pec (see \code{\link[pec]{pec}}).}

\item{...}{Further arguments passet to}

\item{data}{The full data set.}

\item{times}{The time points at which prediction error will be evaluated}

\item{...}{further arguments passed to \code{\link{split_data}}.}
}
\description{
This function is needed to work with the \code{pec} package for estimation 
of the time-dependent C-Index and Brier-Score (prediction error)

k-fold cross validated prediction error curve (pec)

Fit PAM to train data and calculate PEC on test data

Tranform data from pec object to tidy format
}
