% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nest-utils.R
\name{nest_tdc}
\alias{nest_tdc}
\title{Create nested data frame from event and time-dependent data}
\usage{
nest_tdc(event_df, tdc_df, te_var, id_var = "id", time_var = "time",
  status_var = "status", cut = NULL, regular = TRUE, entry_time = 0,
  cens_value = 0, .key = "tdc")
}
\arguments{
\item{event_df}{A data frame (potentially) containing time-dependent covariates.}

\item{tdc_df}{A data frame containing information on time-dependent variables
in long format. Needs to contain a "time" variable indicating when the
TDCs change their value. Must have same name as variable indicating event
times in \code{event_df}.}

\item{te_var}{The time variable in \code{tdc_df} indicating time points at
which time-dependent covariate (tdc) was observed.
Needs to be the same name in both data sets.}

\item{id_var}{The ID variable name, identifying subjects.}

\item{time_var}{A character, specifies the column of the event or
censoring time in \code{event_df} and the time of measurement for
the time-dependent covariates in \code{tdc_df}.}

\item{status_var}{As \code{time_var}, but specifies column containing the
event indicator. Can be missing in the \code{tdc_df}.}

\item{cut}{the vector of timepoints to cut at}

\item{regular}{Logical. Indicates whether covariate data was observed on
a regular grid for all observation units.}

\item{entry_time}{If scalar, the time-point at which the follow up for each
observation unit begins. (Eventually, support for subject specific
entry time could be supported through this argument).}

\item{cens_value}{The value that indicates censoring in the
\code{status_var} column.}

\item{.key}{The name of the new column, as a string or symbol.

This argument is passed by expression and supports
\link[rlang:quasiquotation]{quasiquotation} (you can unquote strings
and symbols). The name is captured from the expression with
\code{\link[rlang:quo_name]{rlang::quo_name()}} (note that this kind of interface where
symbols do not represent actual objects is now discouraged in the
tidyverse; we support it here for backward compatibility).}
}
\description{
Create nested data frame from event and time-dependent data
}
