% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nuclear}
\alias{nuclear}
\title{Time until nuclear power plant construction in different regions.}
\format{
An object of class \code{data.frame} with 724 rows and 3 columns.
}
\usage{
nuclear
}
\description{
This dataset originates from IAEA and contains 730 power.
The data contains the following variables:
\describe{
\item{months}{Construction time}
\item{status}{Event indicator (0 = censored, 1 = construction finished).}
\item{region}{Continent, Africa/Asia, America, Europe, Soviet Union and Warsaw Pact}
}
}
\keyword{datasets}
