% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformData.R
\name{pad_last}
\alias{pad_last}
\title{Add one row to a data frame for step plots.}
\usage{
pad_last(intdf)
}
\arguments{
\item{intdf}{A data frame containing interval (and hazard intformation).}
}
\description{
Add one row to a data frame for step plots.
}

