% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr-wrappers.R, R/pammfit.R
\name{pamm_cr}
\alias{pamm_cr}
\alias{pamm}
\alias{is.pamm}
\alias{print.pamm}
\alias{summary.pamm}
\alias{plot.pamm}
\title{Fit a competing risk piece-wise exponential additive model}
\usage{
pamm_cr(formula, data = list(), method = "REML", ..., engine = "gam")

pamm(
  formula,
  data = list(),
  method = "REML",
  ...,
  trafo_args = NULL,
  engine = "gam"
)

is.pamm(x)

\method{print}{pamm}(x, ...)

\method{summary}{pamm}(object, ...)

\method{plot}{pamm}(x, ...)
}
\arguments{
\item{formula}{A GAM formula, or a list of formulae (see \code{formula.gam}
and also \code{gam.models}). These are exactly like the formula for a GLM
except that smooth terms, \code{s}, \code{te}, \code{ti} and \code{t2}, can
be added to the right hand side to specify that the linear predictor depends
on smooth functions of predictors (or linear functionals of these).}

\item{data}{A PED object (either a \code{ped_cr_union} or \code{ped_cr_list}
object) containing the model status and event time variable such as
covariates required by the formula. By default the variables are taken from
\code{environment(formula)}: typically the environment from which \code{gam}
is called.}

\item{method}{The smoothing parameter estimation method. \code{"GCV.Cp"} to use GCV for unknown scale parameter and
Mallows' Cp/UBRE/AIC for known scale. \code{"GACV.Cp"} is equivalent, but using GACV in place of GCV. \code{"REML"} 
for REML estimation, including of unknown scale, \code{"P-REML"} for REML estimation, but using a Pearson estimate 
of the scale. \code{"ML"} and \code{"P-ML"} are similar, but using maximum likelihood in place of REML. Beyond the 
exponential family \code{"REML"} is the default, and the only other option is \code{"ML"}.}

\item{...}{Further arguments passed to \code{engine}.}

\item{engine}{Character name of the function that will be called to fit the
model. The intended entries are either \code{"gam"} or \code{"bam"}
(both from package \code{mgcv}).}

\item{trafo_args}{A named list. If data is not in PED format, \code{as_ped}
will be called internally with arguments provided in \code{trafo_args}.}

\item{x}{Any R object.}

\item{object}{An object of class \code{pamm} as returned by \code{\link{pamm}}.}
}
\description{
A thin wrapper around \code{\link[mgcv]{gam}},
however, some arguments are prespecified:
\code{family=poisson()}, \code{offset=data$offset} and \code{method="REML"}.
The first two can not be overwritten. The \code{method} argument
can be specified as usual, but defaults to \code{GCV.cp} in
\code{\link[mgcv]{gam}}.
The wrapper works conditionally on the argument \code{data}.
If the competing risks data is supplied in a list, two separate single risk
models are estimated w.r.t. the provided \code{formula}.
If the data is supplied in a joint PED object (\code{ped_cr_union}),
\code{pamm} is straight-forwardly called. Hence, \code{formula} must reflect
how the competing risks are supposed to be modeled. (E.g. as an interaction
term.)
Unlike \code{pamm} this function does not support on-the-spot transformation
to a PED object (via \code{trafo_args}) if no PED object is supplied.

A thin wrapper around \code{\link[mgcv]{gam}}, however, some arguments are
prespecified:
\code{family=poisson()}, \code{offset=data$offset} and \code{method="REML"}.
The first two can not be overwritten. The \code{method} argument
can be specified as usual, but defaults to \code{GCV.cp} in
\code{\link[mgcv]{gam}}.
}
\examples{
data("sir.adm", package = "mvna")
sir_adm <- sir.adm[1:150, ]
ped <- as_ped_cr(
  data         = sir_adm,
  formula      = Surv(time, status) ~ age + pneu
  )
pam <- pamm_cr(ped_status ~ s(tend, by = cause) + age : cause, data = ped)
## Or 
ped <- as_ped_cr(
  data         = sir_adm,
  formula      = Surv(time, status) ~ age + pneu,
  combine      = FALSE
  )
pam <- pamm_cr(ped_status ~ s(tend) + age, data = ped)
ped <- tumor[1:100, ] \%>\%
 as_ped(Surv(days, status) ~ complications, cut = seq(0, 3000, by = 50))
pam <- pamm(ped_status ~ s(tend) + complications, data = ped)
summary(pam)
## Alternatively
pamm(
 ped_status ~ s(tend) + complications,
 data = tumor[1:100, ],
trafo_args = list(formula = Surv(days, status)~complications))
}
\seealso{
\code{\link[mgcv]{gam}}

\code{\link[mgcv]{gam}}
}
\keyword{internal}
