% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr-wrappers.R
\name{pamm_cr.ped_cr_list}
\alias{pamm_cr.ped_cr_list}
\title{Wrapping function for PAMs with competing risks using mgcv::gam}
\usage{
\method{pamm_cr}{ped_cr_list}(
  formula,
  family = poisson(),
  ped = list(),
  offset = NULL,
  bam = FALSE,
  ...
)
}
\arguments{
\item{formula}{an object of class formula or a string convertible to it.
The model formula corresponding to a gam (from the mgcv package).}

\item{family}{the family of the gam to be modelled. Only poisson is
reasonable. Hence, any other input will not be accepted.}

\item{offset}{The offset for each observation. Contained in data.}

\item{...}{additional arguments passed to the gam function.}

\item{data}{A data.frame of class ped_cr that features time-to-event data
and convariates. (see https://adibender.github.io/pammtools/)}
}
\value{
a list of gams - one entry for a single competing risk.
}
\description{
This function serves as a wrapper for the gam function (mgcv) to be
applicable to competing risks. Competing risks are separately modelled
in their own gam. The results can be interpreted synoptically,
using the corresponding methods (summary etc.) or separately using the glm
methods for each entry alone.
The input is entered as for a single risk PAM.
}
\author{
Philipp Kopper
#' library(pammtools)
set.seed(31072019)
df <- cbind.data.frame(
x1 = runif (n, -3, 3),
x2 = runif (n, 0, 6))two component formula specifying cause specific hazards

form <- ~ log(0.5) + x1 + x2 | log(0.9) + (-1) * x1 + (-1 * x2)
df <- sim_pexp_cr(form, df, seq(0, 3, by = 0.25)) \%>\%
mutate(
cens_time = runif(n(), 0.5, 3),
status = if_else(cens_time < time, 0, 1),
time = pmin(time, cens_time),
type = status * type)
df <- df\link{, - c(5, 6)}
colnames(df)\link{7} <- "obs_times"
ped_cr <- as_ped_cr(data = df, Surv(obs_times, status) ~ ., id = "id",
cut = seq(0, max(df$obs_times), 0.25))
pem_cr <- gam_cr(ped_status ~ interval + x1 + x2,
data = ped_cr, offset = offset, family = poisson())
}
