% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr-wrappers.R
\name{pamm_cr}
\alias{pamm_cr}
\alias{pamm_cr.default}
\alias{pamm_cr.ped_cr_list}
\alias{pamm_cr.ped_cr_union}
\title{Fit a competing risk piece-wise exponential additive model}
\usage{
pamm_cr(formula, data = list(), method = "REML", ..., engine = "gam")

\method{pamm_cr}{default}(formula, data = list(), method = "REML", ..., engine = "gam")

\method{pamm_cr}{ped_cr_list}(formula, data = list(), method = "REML", ..., engine = "gam")

\method{pamm_cr}{ped_cr_union}(formula, data = list(), method = "REML", ..., engine = "gam")
}
\arguments{
\item{formula}{A GAM formula, or a list of formulae (see \code{formula.gam}
and also \code{gam.models}). These are exactly like the formula for a GLM
except that smooth terms, \code{s}, \code{te}, \code{ti} and \code{t2}, can
be added to the right hand side to specify that the linear predictor depends
on smooth functions of predictors (or linear functionals of these).}

\item{data}{A PED object (either a \code{ped_cr_union} or \code{ped_cr_list}
object) containing the model status and event time variable such as
covariates required by the formula. By default the variables are taken from
\code{environment(formula)}: typically the environment from which \code{gam}
is called.}

\item{method}{Optimisation method directly handed over to \code{gam}.}

\item{...}{Further arguments passed to \code{engine}.}

\item{engine}{Character name of the function that will be called to fit the
model. The intended entries are either \code{"gam"} or \code{"bam"}
(both from package \code{mgcv}).}
}
\description{
A thin wrapper around \code{\link[mgcv]{gam}},
however, some arguments are prespecified:
\code{family=poisson()}, \code{offset=data$offset} and \code{method="REML"}.
The first two can not be overwritten. The \code{method} argument
can be specified as usual, but defaults to \code{GCV.cp} in
\code{\link[mgcv]{gam}}.
The wrapper works conditionally on the argument \code{data}.
If the competing risks data is supplied in a list, two separate single risk
models are estimated w.r.t. the provided \code{formula}.
If the data is supplied in a joint PED object (\code{ped_cr_union}),
\code{pamm} is straight-forwardly called. Hence, \code{formula} must reflect
how the competing risks are supposed to be modeled. (E.g. as an interaction
term.)
Unlike \code{pamm} this function does not support on-the-spot transformation
to a PED object (via \code{trafo_args}) if no PED object is supplied.
}
\seealso{
\code{\link[mgcv]{gam}}
}
