% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pammtools.R
\docType{package}
\name{pammtools-package}
\alias{pammtools-package}
\title{pammtools: Piece-Wise Exponential Additive Mixed Modeling Tools for Survival Analysis}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

The Piece-wise exponential (Additive Mixed) Model (PAMM; Bender and others (2018) \doi{ 10.1177/1471082X17748083}) is a powerful model class for the analysis of survival (or time-to-event) data, based on Generalized Additive (Mixed) Models (GA(M)Ms). It offers intuitive specification and robust estimation of complex survival models with stratified baseline hazards, random effects, time-varying effects, time-dependent covariates and cumulative effects (Bender and others (2019)), as well as support for left-truncated, competing risks and recurrent events data. pammtools provides tidy workflow for survival analysis with PAMMs, including data simulation, transformation and other functions for data preprocessing and model post-processing as well as visualization.
}
\references{
Bender, Andreas, Andreas Groll, and Fabian Scheipl. 2018.
“A Generalized Additive Model Approach to Time-to-Event Analysis”
Statistical Modelling, February. https://doi.org/10.1177/1471082X17748083.

Bender, Andreas, Fabian Scheipl, Wolfgang Hartl, Andrew G. Day, and Helmut Küchenhoff. 2019.
“Penalized Estimation of Complex, Non-Linear Exposure-Lag-Response Associations.”
Biostatistics 20 (2): 315–31. https://doi.org/10.1093/biostatistics/kxy003.

Bender, Andreas, and Fabian Scheipl. 2018.
“pammtools: Piece-Wise Exponential Additive Mixed Modeling Tools.”
ArXiv:1806.01042 \link{Stat}, June. https://arxiv.org/abs/1806.01042.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://adibender.github.io/pammtools/}
  \item Report bugs at \url{https://github.com/adibender/pammtools/issues}
}

}
\author{
\strong{Maintainer}: Andreas Bender \email{andreas.bender@stat.uni-muenchen.de} (\href{https://orcid.org/0000-0001-5628-8611}{ORCID})

Authors:
\itemize{
  \item Fabian Scheipl \email{fabian.scheipl@stat.uni-muenchen.de} (\href{https://orcid.org/0000-0001-8172-3603}{ORCID})
  \item Philipp Kopper \email{philipp.kopper@stat.uni-muenchen.de} (\href{https://orcid.org/0000-0002-5037-7135}{ORCID})
}

Other contributors:
\itemize{
  \item Lukas Burk \email{burk@leibniz-bips.de} (\href{https://orcid.org/0000-0001-7528-3795}{ORCID}) [contributor]
}

}
