% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pec-helpers.R
\name{pec_pam}
\alias{pec_pam}
\title{Fit PAM to train data and calculate PEC on test data}
\usage{
pec_pam(data, idx_train, idx_test, formula_ped, formula_pam, times = NULL,
  n = 50L, ...)
}
\arguments{
\item{data}{The data in standard time-to-event data format which can be tansformed 
in PED (see \code{\link{split_data}}).}

\item{idx_train}{Index of rows that indicate training set.}

\item{idx_test}{Index of rows that indicate test set.}

\item{formula_ped}{The formula used to transform data to PED (see 
\code{\link{split_data}}).}

\item{formula_pam}{The formula used to fit the PAM (see \code{\link{pam}}).}

\item{times}{The time points at which prediction error will be evaluated}

\item{n}{If \code{times} argument not specified, will create a sequence 
from minimal to maximal time of this length.}

\item{...}{further arguments passed to \code{\link{split_data}}.}
}
\description{
Fit PAM to train data and calculate PEC on test data
}
