% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval-information.R
\name{plot_df}
\alias{plot_df}
\title{Extract information for plotting step functions}
\usage{
plot_df(pinfo)
}
\arguments{
\item{pinfo}{A data frame as returned by \code{\link[pam]{ped_info}} and
potentially additional information from predictions, etc.}
}
\description{
Extract information for plotting step functions
}
\examples{
data("veteran", package="survival")
leuk.ped <- split_data(Surv(time, status)~., data=veteran, id="id")
pem <- glm(ped_status ~ interval, data = leuk.ped, family=poisson(), offset=offset)
pinfo <- ped_info(leuk.ped)
pinfo$basehaz <- predict(pem, newdata=pinfo, type="response")
plot.inf <- plot_df(pinfo)
}
