% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.pamm}
\alias{predict.pamm}
\title{Predict hazard, cumulative hazard or survival probability}
\usage{
\method{predict}{pamm}(object, newdata, type = c("hazard", "cumu_hazard", "surv_prob"), ...)
}
\arguments{
\item{object}{An object of class \code{pam_xgb}}

\item{newdata}{A data set containing the same covariates as used for model
fitting. If of class \code{data.frame}, the function will transform the
data to the PED format using the same settings as for the data used in
estimation.}

\item{type}{The type of prediction desired. Either hazard (\code{type = "hazard"}),
cumulative hazard (\code{type = "cumu_hazard"}) or survival probability
(\code{type = "surv_prob"}).}

\item{...}{Currently not used.}
}
\value{
A matrix of predictions containing one row per
observation (row in newdata) and 1 column per  specified time in the
\code{times} argument.
}
\description{
Predict hazard, cumulative hazard or survival probability
}
\seealso{
pamm
}
