% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predictSurvProb.pamm}
\alias{predictSurvProb.pamm}
\title{S3 method for pamm objects for compatibility with package pec}
\usage{
\method{predictSurvProb}{pamm}(object, newdata, times)
}
\description{
S3 method for pamm objects for compatibility with package pec
}
