% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula-specials.R
\name{prep_concurrent}
\alias{prep_concurrent}
\alias{prep_concurrent.list}
\title{Extract information on concurrent effects}
\usage{
prep_concurrent(x, formula, ...)

\method{prep_concurrent}{list}(x, formula, ...)
}
\arguments{
\item{x}{A suitable object from which variables contained in
\code{formula} can be extracted.}

\item{...}{Further arguments passed to methods.}
}
\description{
Extract information on concurrent effects
}
\keyword{internal}
