% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr-wrappers.R
\name{print.pamm_cr_list}
\alias{print.pamm_cr_list}
\title{Print method for competing risk PAMs (piece-wise additive models)}
\usage{
\method{print}{pamm_cr_list}(summary_list)
}
\arguments{
\item{summary_list}{a list of summaries where each element is one summary
for a \code{pamm}. Each element should correspond to one partial competing
risks model of a PAM(M).}
}
\value{
A (printed) list of summaries.
}
\description{
Print method for competing risk PAMs (piece-wise additive models)
}
\author{
Philipp Kopper
}
