% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-newdata.R
\name{reconstruct_intlen}
\alias{reconstruct_intlen}
\title{Reconstruct intlen from sorted unique time values}
\usage{
reconstruct_intlen(newdata, time_var = "tend")
}
\arguments{
\item{newdata}{A data frame with a time column.}

\item{time_var}{Character string naming the time variable in
\code{newdata}. Defaults to \code{"tend"}.}
}
\value{
The input data frame with an \code{intlen} column added.
}
\description{
Computes interval lengths from the sorted unique values of the time variable.
This is used by add_* functions that need intlen for cumulative calculations.
}
\keyword{internal}
