% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group-helpers.R
\name{rm_grpvars}
\alias{rm_grpvars}
\title{Return ungrouped data frame without grouping variables}
\usage{
rm_grpvars(data)
}
\arguments{
\item{data}{A data frame (potentially `grouped_df`) from which grouping 
variables will be removed.}
}
\value{
Returns \code{data} without grouping variables (and group properties).
}
\description{
Return ungrouped data frame without grouping variables
}
