% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpexp.R
\name{rpexp}
\alias{rpexp}
\title{Draw random numbers from piece-wise exponential distribution.}
\usage{
rpexp(n = 1, rate = 1, t = 0)
}
\description{
This is a copy of the same function from \code{rpexp} from package
\pkg{msm}.
Copied here to reduce dependencies.
}
\keyword{internal}
