% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct-newdata.R
\name{seq_range}
\alias{seq_range}
\title{Creates sequence from minimum to maximum}
\usage{
seq_range(x, length.out = 100L)
}
\arguments{
\item{x}{A numeric or integer vector.}

\item{length.out}{desired length of the sequence.  A
    non-negative number, which for \code{seq} and \code{seq.int} will be
    rounded up if fractional.}
}
\description{
Creates sequence from minimum to maximum
}
