% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{shiny_viz}
\alias{shiny_viz}
\title{Shiny visualization of an estimated model}
\usage{
shiny_viz(data, model, term, reference = NULL, ci = TRUE,
  launch.browser = TRUE, ...)
}
\arguments{
\item{model}{Estimated model object}

\item{reference}{If specified, should be a list with covariate value pairs,
e.g. \code{list(x1 = 1, x=50)}. The calculated partial effect will be relative
to an observation specified in \code{reference}.}

\item{launch.browser}{Run the shiny app in the browser}

\item{...}{additional arguments, passed to \code{\link{gg_partial}},
\code{\link{gg_partial_ll}} and \code{\link{gg_slice}}}
}
\description{
Based on an estimated model, start a shiny-based interactive
visualization of the effect structures.
}
