% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation-functions.R
\name{sim_wce}
\alias{sim_wce}
\alias{sim_wce2}
\title{Simulate survival data using the piece-wise exponential distribution}
\usage{
sim_wce(Xdf, tmax = 30)

sim_wce2(Xdf, tmax = 30)
}
\arguments{
\item{Xdf}{A data frame containing full (time-dependent) covariate information
for each subject}

\item{tmax}{The maximal time of follow-up. Survival times after \code{tmax}
will be administratively censored.}
}
\description{
Simulate survival data using the piece-wise exponential distribution
}
\keyword{internal}
