% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simdf_elra}
\alias{simdf_elra}
\title{Simulated data with cumulative effects}
\format{An object of class \code{ped} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 250 rows and 9 columns.}
\usage{
simdf_elra
}
\description{
Simulated data with cumulative effects
}
\keyword{datasets}
