% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penalized-lag-lead.R
\name{smooth.construct.fdl.smooth.spec}
\alias{smooth.construct.fdl.smooth.spec}
\title{New basis for penalized lag selection}
\usage{
\method{smooth.construct}{fdl.smooth.spec}(object, data, knots)
}
\arguments{
\item{object}{An object handled by mgcv}

\item{data}{The data set}

\item{knots}{A vector of knots}
}
\description{
Originally proposed in  Obermeier et al., 2015, Flexible Distributed Lags for Modelling Earthquake Data,
Journal of the Royal Statistical Society: Series C (Applied Statistics),
10.1111/rssc.12077.
Here extended in order to penalize lead times in addition to lag times.
Ideally the lag-lead window would then be selected in a data-driven fashion.
Treat as experimental.
}
\keyword{internal}
