% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-terms.R
\name{smooth_term_rows}
\alias{smooth_term_rows}
\title{Turn a single mgcv smooth into zero or more curve specifications}
\usage{
smooth_term_rows(s, data)
}
\arguments{
\item{s}{A single smooth object from \code{fit$smooth}.}

\item{data}{A data frame containing the variables used to fit the model.}
}
\description{
Turn a single mgcv smooth into zero or more curve specifications
}
\keyword{internal}
