% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula-specials.R
\name{func}
\alias{func}
\alias{concurrent}
\title{Formula specials for defining  time-dependent covariates}
\usage{
func(..., te_var, ll_fun = function(t, te) {     t >= te }, suffix = NULL)

concurrent(..., te_var, ll_fun = function(t) {     t == t }, suffix = NULL)
}
\description{
So far, two specials are implemented. \code{concurrent} is used when
the goal is to estimate a concurrent effect of the TDC. \code{func}
is used when the goal is to estimate a cumulative effect of the TDC.
}
\keyword{internal}
