% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr-wrappers.R
\name{summary.pamm_cr_list}
\alias{summary.pamm_cr_list}
\title{Summary method for competing risk PAMs (piece-wise additive models)}
\usage{
\method{summary}{pamm_cr_list}(pam_cr)
}
\arguments{
\item{pem_cr}{An object of class pem_cr where all elements are one
glm object. Each element should correspond to one partial competing risks
model of a PEM.}
}
\value{
A list of summaries.
}
\description{
This function summarises the underlying models of a pem_cr object.
The summaries are returned list-wise with each element belonging to
one competing risk.
}
\author{
Philipp Kopper
}
