% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidiers.R
\name{tidy_fixed}
\alias{tidy_fixed}
\title{Extract fixed coefficient table from model object}
\usage{
tidy_fixed(x, ...)
}
\arguments{
\item{x}{A model object.}

\item{...}{Currently not used.}
}
\description{
Given a model object, returns data frame with columns \code{variable},
\code{coef} (coefficient), \code{lower} (lower 95\% CI) and
\code{upper} (upper 95% CI).
}
