% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pec-helpers.R
\name{tidy_pec}
\alias{tidy_pec}
\title{Tranform data from pec object to tidy format}
\usage{
tidy_pec(pec)
}
\arguments{
\item{pec}{An object of class pec (see \code{\link[pec]{pec}}).}
}
\description{
Tranform data from pec object to tidy format
}
