% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidiers.R
\name{tidy_re}
\alias{tidy_re}
\title{Extract random effects objects in tidy data format.}
\usage{
tidy_re(x, keep = c("fit", "main", "xlab", "ylab"), ...)
}
\arguments{
\item{x}{A model object.}

\item{keep}{A vector of variables to keep.}

\item{...}{Currently not used.}
}
\description{
Extract random effects objects in tidy data format.
}
\seealso{
\code{\link[stats]{qqline}}
}
