% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unsplit.R
\name{unsplit}
\alias{unsplit}
\title{Transform PED data back to classic format}
\usage{
unsplit(ped, id_var = "id")
}
\arguments{
\item{ped}{A data set in PED format}

\item{id_var}{Name of the id column}
}
\description{
Transform PED data back to classic format
}
\keyword{internal}
