% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warnings.R
\name{warn_about_new_time_points.glm}
\alias{warn_about_new_time_points.glm}
\title{Warn if new t_j are used}
\usage{
\method{warn_about_new_time_points}{glm}(object, newdata, time_var, ...)
}
\description{
Warn if new t_j are used
}
\keyword{internal}
