% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warnings-and-errors.R
\name{warn_about_new_time_points}
\alias{warn_about_new_time_points}
\title{Warn if new t_j are used}
\usage{
warn_about_new_time_points(newdata, object, time_var)
}
\description{
Warn if new t_j are used
}
\keyword{internal}
