% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-functions.R, R/pool-ic.R
\name{add_cif}
\alias{add_cif}
\alias{add_cif.default}
\alias{add_cif.pamm_ic}
\title{Add cumulative incidence function to data}
\usage{
add_cif(newdata, object, ...)

\method{add_cif}{default}(
  newdata,
  object,
  ci = TRUE,
  overwrite = FALSE,
  alpha = 0.05,
  nsim = 500L,
  cause_var = "cause",
  time_var = NULL,
  interval_length = "intlen",
  ...
)

\method{add_cif}{pamm_ic}(
  newdata,
  object,
  ci = TRUE,
  alpha = 0.05,
  nsim = 500L,
  cause_var = "cause",
  time_var = NULL,
  interval_length = "intlen",
  ...
)
}
\arguments{
\item{newdata}{ A data frame or list containing the values of the model covariates at which predictions
  are required. If this is not provided then predictions corresponding to the
                original data are returned. If \code{newdata} is provided then
                it should contain all the variables needed for prediction: a
                warning is generated if not. See details for use with \code{link{linear.functional.terms}}. }

\item{object}{ a fitted \code{gam} object as produced by \code{gam()}.
                }

\item{...}{Further arguments passed to \code{\link[mgcv]{predict.gam}} and
\code{\link{get_hazard}}}

\item{ci}{\code{logical}. Indicates if confidence intervals should be
calculated. Defaults to \code{TRUE}.}

\item{overwrite}{Should hazard columns be overwritten if already present in
the data set? Defaults to \code{FALSE}. If \code{TRUE}, columns with names
\code{c("hazard", "se", "lower", "upper")} will be overwritten.}

\item{alpha}{Significance level for pooled confidence intervals.}

\item{nsim}{Total number of pooled posterior draws used for the interval.}

\item{cause_var}{Character. Column name of the 'cause' variable.}

\item{time_var}{Name of the variable used for the baseline hazard. Defaults
to \code{"tend"}.}

\item{interval_length}{\code{Character}, defaults to \code{"intlen"}.
contains the interval length in \code{newdata}.}
}
\description{
Add cumulative incidence function to data
}
\details{
When computing cumulative incidence for multiple groups, the input data must
be grouped via \code{group_by()} before calling this function. Omitting
\code{group_by()} will not produce an error or warning but will return
silently incorrect results, as the cumulative incidence will be accumulated
over the entire dataset rather than within each group.
}
\examples{
\donttest{
if (require("etm")) {
  data("fourD", package = "etm")
  ped_stacked <- fourD |>
    dplyr::select(-medication, -treated) |>
    as_ped(Surv(time, status) ~., id = "id") |>
    dplyr::mutate(cause = as.factor(cause))
  pam <- pamm(
    ped_status ~ s(tend, by = cause) + sex + sex:cause + age + age:cause,
    data = ped_stacked)
  ped_stacked |>
    make_newdata(tend = unique(tend), cause = unique(cause)) |>
    group_by(cause) |>
    add_cif(pam)
}
}
   
}
