% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hazard.R
\name{add_hazard}
\alias{add_cumhazard}
\alias{add_hazard}
\title{Add predicted hazard to data set}
\usage{
add_hazard(newdata, object, se.fit = TRUE, type = c("response", "link"),
  ...)

add_cumhazard(newdata, object, se.fit = TRUE, type = "response", ...)
}
\arguments{
\item{newdata}{ A data frame or list containing the values of the model covariates at which predictions
  are required. If this is not provided then predictions corresponding to the
                original data are returned. If \code{newdata} is provided then
                it should contain all the variables needed for prediction: a
                warning is generated if not. }

\item{object}{ a fitted \code{gam} object as produced by \code{gam()}.
                }

\item{se.fit}{ when this is TRUE (not default) standard error estimates are returned for each prediction.}

\item{type}{ When this has the value \code{"link"} (default) the linear predictor (possibly with
associated standard errors) is returned. When \code{type="terms"} each component of the 
linear predictor is returned seperately (possibly with standard errors): this includes 
parametric model components, followed by each smooth component, but excludes
any offset and any intercept. \code{type="iterms"} is the same, except that any standard errors 
returned for smooth components will include the uncertainty about the intercept/overall mean.  When 
\code{type="response"} predictions 
on the scale of the response are returned (possibly with approximate
standard errors). When \code{type="lpmatrix"} then a matrix is returned
which yields the values of the linear predictor (minus any offset) when
postmultiplied by the
parameter vector (in this case \code{se.fit} is ignored). The latter
option is most useful for getting variance estimates for quantities derived from
the model: for example integrated quantities, or derivatives of smooths. A
linear predictor matrix can also be used to implement approximate prediction
outside \code{R} (see example code, below). }

\item{...}{Further arguments passed to \code{\link[mgcv]{predict.gam}}}
}
\description{
Add predicted hazard to data set

Add cumulative hazard estimate to data set
}
\examples{
library(mgcv)
data("leuk2", package="bpcp")
leuk.ped <- split_data(Surv(time, status)~., data=leuk2, id="id")
pam <- gam(status ~ s(tend), data = leuk.ped, family=poisson(), offset=offset)
pinfo <- ped_info(leuk.ped)
add_hazard(pinfo, pam)
library(mgcv)
data("leuk2", package="bpcp")
leuk.ped <- split_data(Surv(time, status)~., data=leuk2, id="id")
pam <- gam(status ~ s(tend), data = leuk.ped, family=poisson(), offset=offset)
pinfo <- ped_info(leuk.ped)
add_cumhazard(pinfo, pam)
}
\seealso{
\code{\link[mgcv]{predict.gam}}, \code{\link[pam]{add_cumhazard}}

\code{\link[mgcv]{predict.gam}}, \code{\link[pam]{add_hazard}}
}

