% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hazard.R
\name{add_hazard}
\alias{add_hazard}
\alias{add_cumhazard}
\title{Add predicted hazard to data set}
\usage{
add_hazard(newdata, object, type = c("response", "link"), ci = TRUE,
  se.mult = 2, overwrite = FALSE, time_variable = NULL, ...)

add_cumhazard(newdata, object, ci = TRUE, se.mult = 2, overwrite = FALSE,
  time_variable = NULL, interval_length = quo(intlen), ...)
}
\arguments{
\item{newdata}{ A data frame or list containing the values of the model covariates at which predictions
  are required. If this is not provided then predictions corresponding to the
                original data are returned. If \code{newdata} is provided then
                it should contain all the variables needed for prediction: a
                warning is generated if not. See details for use with \code{link{linear.functional.terms}}. }

\item{object}{ a fitted \code{gam} object as produced by \code{gam()}.
                }

\item{type}{ When this has the value \code{"link"} (default) the linear predictor (possibly with
associated standard errors) is returned. When \code{type="terms"} each component of the 
linear predictor is returned seperately (possibly with standard errors): this includes 
parametric model components, followed by each smooth component, but excludes
any offset and any intercept. \code{type="iterms"} is the same, except that any standard errors 
returned for smooth components will include the uncertainty about the intercept/overall mean.  When 
\code{type="response"} predictions 
on the scale of the response are returned (possibly with approximate
standard errors). When \code{type="lpmatrix"} then a matrix is returned
which yields the values of the linear predictor (minus any offset) when
postmultiplied by the
parameter vector (in this case \code{se.fit} is ignored). The latter
option is most useful for getting variance estimates for quantities derived from
the model: for example integrated quantities, or derivatives of smooths. A
linear predictor matrix can also be used to implement approximate prediction
outside \code{R} (see example code, below). }

\item{ci}{Logical indicating whether to iclude confidence intervals. Defaults
to \code{TRUE}}

\item{se.mult}{Factor by which standard errors are multiplied for calculating
the confidence intervals.}

\item{overwrite}{Should hazard columns be overwritten if already present in
the data set? Defaults to \code{FALSE}. If \code{TRUE}, columns with names
\code{c("hazard", "se", "lower", "upper")} will be overwritten.}

\item{time_variable}{Name of the variable used for the baseline hazard. If
not given, defaults to \code{"tend"} for \code{\link[mgcv]{gam}} fits, else
\code{"interval"}. The latter is assumed to be a factor, the former
numeric.}

\item{...}{Further arguments passed to \code{\link[mgcv]{predict.gam}} and
\code{\link{get_hazard}}}

\item{interval_length}{\code{quosure} providing the name of the variable in
newdata containing the interval lengths. Defaults to \code{intlen}.}
}
\description{
Add predicted hazard to data set

Add cumulative hazard estimate to data set
}
\examples{
\dontrun{
library(mgcv)
data("veteran", package="survival")
ped <- split_data(Surv(time, status)~. cut=seq(0, 500, by=100), data=veteran, 
	id="id")
pam <- gam(ped_status ~ s(tend, k=5), data = ped, family=poisson(), offset=offset)
pinfo <- ped_info(ped)
add_hazard(pinfo, pam)
}
}
\seealso{
\code{\link[mgcv]{predict.gam}}, \code{\link[pam]{add_cumhazard}}

\code{\link[mgcv]{predict.gam}}, \code{\link[pam]{add_hazard}}
}
