% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-functions.R
\name{add_trans_ci}
\alias{add_trans_ci}
\title{Add transition probabilities confidence intervals}
\usage{
add_trans_ci(newdata, object, n_sim = 100L, alpha = 0.05, ...)
}
\description{
Add transition probabilities confidence intervals
}
\keyword{internal}
