% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-functions.R
\name{add_trans_prob}
\alias{add_trans_prob}
\title{Add transition probabilities}
\usage{
add_trans_prob(
  newdata,
  object,
  overwrite = FALSE,
  ci = FALSE,
  alpha = 0.05,
  n_sim = 100L,
  time_var = NULL,
  interval_length = "intlen",
  ...
)
}
\description{
Add transition probabilities
}
