% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-ped.R
\name{as_ped}
\alias{as_ped}
\alias{is.ped}
\title{Transform data to Piece-wise Exponential Data (PED)}
\usage{
as_ped(data, formula, ...)

is.ped(x)
}
\arguments{
\item{data}{a data frame}

\item{formula}{a model formula}

\item{...}{Further arguments passed to \code{\link[survival]{survSplit}}}

\item{x}{any R object.}
}
\value{
A data frame class \code{ped} in piece-wise exponential data format.
}
\description{
Transform data to Piece-wise Exponential Data (PED)
}
\seealso{
\code{\link[survival]{survSplit}}
}
