% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-trafo.R
\name{as_ped}
\alias{as_ped}
\alias{is.ped}
\title{Function to check if an object is of class \code{ped} or 
transform data into piece-wise exponential data (PED) format if possible.}
\usage{
as_ped(x, data, tdc_df = NULL, ...)

is.ped(x)
}
\arguments{
\item{x}{any R object.}

\item{data}{a data frame}

\item{tdc_df}{A data set containing information on time-dependent covariates (TDC).}

\item{...}{Further arguments passed to methods.}
}
\value{
A data frame class \code{ped} in piece-wise exponential data (PED) format.
}
\description{
Function to check if an object is of class \code{ped} or 
transform data into piece-wise exponential data (PED) format if possible.
}
