% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr.R
\name{as_ped_cr}
\alias{as_ped_cr}
\title{Transform data to Piece-wise Exponential Data (PED) for a Cause-Specific
hazards competing risks model}
\usage{
as_ped_cr(
  data,
  formula,
  censor_code = 0L,
  output = c("data.frame", "list"),
  cut = NULL,
  ...
)
}
\arguments{
\item{data}{Either an object inheriting from data frame or in case of
time-dependent covariates a list of data frames, where the first data frame
contains the time-to-event information and static covariates while the
second (and potentially further data frames) contain information on
time-dependent covariates and the times at which they have been observed.}

\item{formula}{A two sided formula with a Surv object on the left-hand-side
and covariate specification on the right-hand-side (RHS). The RHS can be an
extended formula, which specifies how TDCs should be transformed using
specials concurrent and cumulative.}

\item{censor_code}{Either a string or integer (depending on data) with out
outlines which level in the status variable is associated with censorship.}

\item{output}{A chracter value: either "data.frame" or "list".
Will be matched via match.arg().}

\item{cut}{Break points, used to partition the follow up into intervals.
If unspecified, all unique event times will be used.}

\item{...}{Further arguments passed to the data.frame method and possibly
to survSplit}
}
\value{
A list of PED data.frames.
}
\description{
This is a wrapper for the general data transformation function provided by
the pammtools package (as_ped). The use is identical to as_ped.
However, the function creates one single PED object for each risk.
The results are stored in a list
For more details use ?as_ped.
For each risk, all competing risks are assumed to be right-censored.
}
\author{
Philipp Kopper
}
