% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct-newdata.R
\name{combine_df}
\alias{combine_df}
\title{Combines multiple data frames}
\usage{
combine_df(...)
}
\arguments{
\item{...}{Data frames that should be combined to one data frame. 
Elements of first df vary fastest, elements of last df vary slowest.}
}
\description{
Works like \code{\link[base]{expand.grid}} but for data frames.
}
\examples{
combine_df(
  data.frame(x=1:3, y=3:1), 
  data.frame(x1=c("a", "b"), x2=c("c", "d")),
  data.frame(z=c(0, 1)))
}
