% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedMethods.R
\name{dplyr_verbs}
\alias{dplyr_verbs}
\alias{filter.ped}
\alias{filter_.ped}
\alias{slice.ped}
\alias{slice_.ped}
\alias{arrange.ped}
\alias{arrange_.ped}
\alias{select_.ped}
\alias{rename_.ped}
\alias{distinct_.ped}
\alias{mutate.ped}
\alias{mutate_.ped}
\alias{transmute.ped}
\alias{transmute_.ped}
\alias{summarise_.ped}
\alias{sample_n.ped}
\alias{sample_frac.ped}
\alias{left_join.ped}
\alias{group_by_.ped}
\alias{ungroup.ped}
\title{\code{dplyr} Verbs for \code{ped}-Objects}
\usage{
filter.ped(.data, ...)

\method{filter_}{ped}(.data, ...)

slice.ped(.data, ...)

\method{slice_}{ped}(.data, ...)

arrange.ped(.data, ...)

\method{arrange_}{ped}(.data, ...)

\method{select_}{ped}(.data, ...)

\method{rename_}{ped}(.data, ...)

\method{distinct_}{ped}(.data, ...)

mutate.ped(.data, keep.attributes = TRUE, ...)

\method{mutate_}{ped}(.data, keep.attributes = TRUE, ...)

transmute.ped(.data, ...)

\method{transmute_}{ped}(.data, ...)

\method{summarise_}{ped}(.data, ...)

\method{sample_n}{ped}(.data, ...)

\method{sample_frac}{ped}(.data, ...)

\method{left_join}{ped}(x, y, by = NULL, copy = FALSE, suffix = c(".x",
  ".y"), ...)

\method{group_by_}{ped}(.data, ...)

\method{ungroup}{ped}(.data, ...)
}
\arguments{
\item{.data}{an  object of class \code{ped}, see \code{\link{split_data}}.}

\item{...}{see \code{dplyr} documentation}

\item{keep.attributes}{conserve attributes? defaults to \code{TRUE}}

\item{x}{tbls to join}

\item{y}{tbls to join}

\item{by}{a character vector of variables to join by.  If \code{NULL}, the
  default, \code{join} will do a natural join, using all variables with
  common names across the two tables. A message lists the variables so
  that you can check they're right (to suppress the message, simply
  explicitly list the variables that you want to join).

  To join by different variables on x and y use a named vector.
  For example, \code{by = c("a" = "b")} will match \code{x.a} to
  \code{y.b}.}

\item{copy}{If \code{x} and \code{y} are not from the same data source,
and \code{copy} is \code{TRUE}, then \code{y} will be copied into the
same src as \code{x}.  This allows you to join tables across srcs, but
it is a potentially expensive operation so you must opt into it.}

\item{suffix}{If there are non-joined duplicate variables in \code{x} and
\code{y}, these suffixes will be added to the output to diambiguate them.}
}
\value{
a modified \code{ped} object
}
\description{
See \code{dplyr} documentation of the respective function for
  details and examples.
}
