% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func-formula.R
\name{func}
\alias{func}
\alias{has_latency}
\title{A formula special for defining functional covariates}
\usage{
func(..., by = NULL, ll_fun = function(t, te) {     te <= t })

has_latency(vars)
}
\description{
A formula special for defining functional covariates
}
\keyword{internal}
