% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_hazard.R
\docType{data}
\name{geom_hazard}
\alias{geom_hazard}
\alias{GeomHazard}
\alias{geom_stephazard}
\alias{GeomStepHazard}
\alias{geom_surv}
\alias{GeomSurv}
\title{PAMM (Cumulative) (Step-)Hazard Plots.}
\usage{
geom_hazard(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)

geom_stephazard(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", direction = "vh", na.rm = FALSE,
  show.legend = NA, inherit.aes = TRUE, ...)

geom_surv(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\description{
\code{geom_hazard} is an extension of the \code{geom_line}, and
is optimized for (cumulative) hazard plots. Essentially, it add a (0,0)
row to the data, if not already the case.
}
\section{Aesthetics}{

\Sexpr[results=rd,stage=build]{ggplot2:::rd_aesthetics("geom", "line")}
}

\examples{
library(ggplot2)
library(pammtools)
ped <- tumor[10:50,] \%>\% as_ped(Surv(days, status)~1)
pam <- pamm(ped_status ~ s(tend), data=ped)
ndf <- make_newdata(ped, tend = unique(tend)) \%>\% add_hazard(pam)
# piece-wise constant hazards
ggplot(ndf, aes(x = tend, y = hazard)) +
 geom_vline(xintercept = c(0, ndf$tend[c(1, (nrow(ndf)-2):nrow(ndf))]), lty = 3) +
 geom_hline(yintercept = c(ndf$hazard[1:3], ndf$hazard[nrow(ndf)]), lty = 3) +
 geom_stephazard() +
 geom_step(col=2) +
 geom_step(col=2, lty = 2, direction="vh")

# comulative hazard
ndf <- ndf \%>\% add_cumu_hazard(pam)
ggplot(ndf, aes(x = tend, y = cumu_hazard)) +
 geom_hazard() +
 geom_line(col=2) # doesn't start at (0, 0)

# survival probability
ndf <- ndf \%>\% add_surv_prob(pam)
ggplot(ndf, aes(x = tend, y = surv_prob)) +
 geom_surv() +
 geom_line(col=2) # doesn't start at c(0,1)
}
\seealso{
\code{\link[ggplot2]{geom_line}},
  \code{\link[ggplot2]{geom_step}}.
}
\keyword{datasets}
