% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-extensions.R
\docType{data}
\name{geom_stepribbon}
\alias{geom_stepribbon}
\alias{GeomStepribbon}
\title{Step ribbon plots.}
\usage{
geom_stepribbon(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  direction = "hv",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\description{
\code{geom_stepribbon} is an extension of the \code{geom_ribbon}, and
is optimized for Kaplan-Meier plots with pointwise confidence intervals
or a confidence band. The default \code{direction}-argument \code{"hv"} is
appropriate for right-continuous step functions like the hazard rates etc
returned by \code{pammtools}.
}
\examples{
library(ggplot2)
huron <- data.frame(year = 1875:1972, level = as.vector(LakeHuron))
h <- ggplot(huron, aes(year))
h + geom_stepribbon(aes(ymin = level - 1, ymax = level + 1), fill = "grey70") +
    geom_step(aes(y = level))
h + geom_ribbon(aes(ymin = level - 1, ymax = level + 1), fill = "grey70") +
    geom_line(aes(y = level))
}
\seealso{
\code{\link[ggplot2]{geom_ribbon}} \code{geom_stepribbon}
inherits from \code{geom_ribbon}.
}
\keyword{datasets}
