% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-functions.R
\name{get_cif}
\alias{get_cif}
\alias{get_cif.default}
\title{Calculate CIF for one cause}
\usage{
get_cif(newdata, object, ...)

\method{get_cif}{default}(
  newdata,
  object,
  ci,
  time_var,
  alpha,
  n_sim,
  cause_var,
  coefs,
  V,
  sim_coef_mat,
  ...
)
}
\description{
Calculate CIF for one cause
}
\keyword{internal}
