% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumulative-coefficient.R
\name{get_cumu_coef_baseline}
\alias{get_cumu_coef_baseline}
\title{Extract cumulative coefficients (cumulative hazard differences)}
\usage{
get_cumu_coef_baseline(data, model, ...)
}
\arguments{
\item{data}{Additional data if necessary.}

\item{model}{Object from which to extract cumulative coefficients.}

\item{...}{Further arguments passed to methods.}
}
\description{
These functions are designed to extract (or mimic) the cumulative coefficients
usually used in additive hazards models (Aalen model) to depict (time-varying)
covariate effects. For PAMMs, these are the differences
between the cumulative hazard rates where all covariates except one have the
identical values. For a numeric covariate of interest, this calculates
\eqn{\Lambda(t|x+1) - \Lambda(t|x)}.  For non-numeric covariates
the cumulative hazard of the reference level is subtracted from
the cumulative hazards evaluated at all non reference levels. Standard
errors are calculated using the delta method.
}
\keyword{internal}
