% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-cut-points.R
\name{get_cut}
\alias{get_cut}
\alias{get_cut.default}
\alias{get_cut.list}
\title{Obtain interval break points}
\usage{
get_cut(data, formula, cut = NULL, ...)

\method{get_cut}{default}(data, formula, cut = NULL, max_time = NULL, event = 1L, ...)

\method{get_cut}{list}(
  data,
  formula,
  cut = NULL,
  max_time = NULL,
  event = 1L,
  timescale = "gap",
  ...
)
}
\description{
Default method words for data frames.
The list method applies the default method to each data set within the list.
}
\keyword{internal}
