% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trafo-cumulative.R
\name{get_func}
\alias{get_func}
\alias{expand_func}
\title{Expand time-dependent covariates to functionals}
\usage{
get_func(data, formula)

expand_func(data, func)
}
\arguments{
\item{data}{Data frame (or similar) in which variables specified in ...
will be looked for}

\item{func}{Single \code{\link{func}} term \code{\link{get_func}} \code{formula}
evaluated.}

\item{...}{One sided formula of the form \code{~func(t,te,x) + func(t-te, z)}
that specifies the type of cumulative effect desired (see examples).
Currently it is assumed that observations of covariates \code{x}, \code{z}
happened on the same time scale \code{te}.
Possible specifications are
\itemize{
\item \code{func(t,te,x)}: The most general specification.
\item \code{func(t, t-te,x)}: A time-varying DLNM.
\item \code{func(t-te,x}: A DLNM.
\item \code{func(t-te,by=x) A WCE}}}
}
\description{
Given formula specification on how time-dependent covariates affect the
outcome, creates respective functional covariate as well as auxiliary
matrices for time/latency etc.
}
\keyword{internal}
