% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz-elra.R
\name{get_partial_ll}
\alias{get_partial_ll}
\title{Visualize effect estimates for specific covariate combinations}
\usage{
get_partial_ll(x, model, term, ..., reference = NULL, ci = FALSE)
}
\arguments{
\item{model}{A suitable model object which will be used to estimate the
partial effect of \code{term}.}

\item{...}{Covariate specifications (expressions) that will be evaulated
by looking for variables in \code{x}. Must be of form \code{z = f(z)}
where \code{z} is a variable in the data set \code{x} and \code{f} a known
function that can be usefully applied to \code{age}. See examples below.}

\item{reference}{If specified, should be a list with covariate value pairs,
e.g. \code{list(x1 = 1, x=50)}. The calculated partial effect will be relative
to an observation specified in \code{reference}.}
}
\description{
Depending on the plot function and input, creates either a 1-dimensional slices,
bivariate surface or  (1D) cumulative effect.
}
\keyword{internal}
