% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-functions.R
\name{get_sim_ci}
\alias{get_sim_ci}
\title{Calculate simulation based confidence intervals}
\usage{
get_sim_ci(newdata, object, alpha = 0.05, nsim = 100L, ...)
}
\description{
Calculate simulation based confidence intervals
}
\keyword{internal}
