% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-terms.R
\name{get_term}
\alias{get_term}
\title{Extract partial effects for specified model terms}
\usage{
get_term(data, fit, term, n = 100, ...)
}
\arguments{
\item{data}{A data frame containing variables used to fit the model. Only
first row will be used.}

\item{fit}{A fitted object of class \code{\link[mgcv]{gam}}.}

\item{term}{The (non-linear) model term of interest.}

\item{n}{Specify the output sequence either by supplying the
length of the sequence with \code{n}, or the spacing between value
with \code{by}. Specifying both is an error.

I recommend that you name these arguments in order to make it clear to
the reader.}

\item{...}{Further arguments passed to \code{\link{seq_range}}.}
}
\description{
Extract partial effects for specified model terms
}
\keyword{internal}
