% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-terms.R
\name{get_terms}
\alias{get_terms}
\title{Extract the partial effects of non-linear model terms}
\usage{
get_terms(data, fit, terms, ...)
}
\arguments{
\item{data}{Any data frame containing variables used to fit the model. Only
first row will be used.}

\item{fit}{A fitted object of class \code{\link[mgcv]{gam}}.}

\item{terms}{A character vector (can be length one). Specifies the terms
for which partial effects will be returned}

\item{...}{Currently ignored.}
}
\value{
A data frame with 5 columns.
}
\description{
Extract the partial effects of non-linear model terms
}
\examples{
library(survival)
fit <- coxph(Surv(time, status) ~ pspline(karno, df=4), data=veteran)
term.karno <- get_terms(veteran, fit, terms="karno")
}
\seealso{
\code{\link[survival]{coxph}}
}
