% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience-plots.R
\name{gg_fixed}
\alias{gg_fixed}
\title{Forrest plot of fixed coefficients}
\usage{
gg_fixed(x, intercept = FALSE, ...)
}
\arguments{
\item{x}{A model object.}

\item{intercept}{Should intercept also be returned?}

\item{...}{Currently not used.}
}
\description{
Forrest plot of fixed coefficients
}
\examples{
g <- mgcv::gam(Sepal.Length ~ Sepal.Width + Petal.Length + Petal.Width + Species, 
	data=iris)
gg_fixed(g, intercept=TRUE)
gg_fixed(g)
}
\seealso{
\code{\link{tidy_fixed}}
}
