% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz-elra.R
\name{gg_partial}
\alias{gg_partial}
\alias{gg_slice}
\title{Visualize effect estimates for specific covariate combinations}
\usage{
gg_partial(data, mod, term, ..., reference = NULL)

gg_slice(data, mod, term, ..., reference = NULL)
}
\description{
Depending on the plot type and input, creates either a bivariate effect
surface or 1-dimensional slices.
}
\keyword{internal}
\keyword{internaldo}
