% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conveniencePlots.R
\name{gg_smooth}
\alias{gg_smooth}
\title{Plot smooth terms of gam objects}
\usage{
gg_smooth(data, fit, ...)
}
\value{
A \code{\link[ggplot2]{ggplot2}} object.
}
\description{
Given a gam model this convenience functions returns a plot of all 
smooth terms contained in the model. If more than one smooth is present, the 
different smooth are faceted.
}
\seealso{
\code{\link[mgcvtools]{tidy_smooth}}
}

