% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conveniencePlots.R
\name{gg_smooth}
\alias{gg_smooth}
\title{Plot smooth 1d terms of gam objects}
\usage{
gg_smooth(data, fit, ...)
}
\arguments{
\item{data}{Any data frame containing variables used to fit the model. Only
first row will be used.}

\item{fit}{A fitted object of class \code{gam}.}

\item{...}{Further arguments passed to \code{\link{get_terms}}}
}
\value{
A \code{\link[ggplot2]{ggplot2}} object.
}
\description{
Given a gam model this convenience function returns a plot of all 
smooth terms contained in the model. If more than one smooth is present, the 
different smooth are faceted.
}
\examples{
g1 <- mgcv::gam(Sepal.Length ~ s(Sepal.Width) + s(Petal.Length), data=iris)
gg_smooth(iris, g1, terms=c("Sepal.Width", "Petal.Length"))
}

