% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdc-utils.R
\name{has_tdc}
\alias{has_tdc}
\title{Checks if data contains timd-dependent covariates}
\usage{
has_tdc(data, id_var)
}
\arguments{
\item{data}{A data frame (potentially) containing time-dependent covariates.}

\item{id_var}{A character indicating the grouping variable. For each covariate
it will be checked if their values change within a group specified by
\code{id_var}.}
}
\value{
Logical. \code{TRUE} if data contains time-dependent covariates, else \code{FALSE}.
}
\description{
Checks if data contains timd-dependent covariates
}
\keyword{internal}
