% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval_info.R
\name{int_info}
\alias{int_info}
\title{Given breaks, create start/end times and interval information}
\usage{
int_info(brks, min.time = 0L)
}
\arguments{
\item{brks}{numeric. A vector of cut point in which the follow-up should be
partitioned in.}

\item{min.time}{numeric. Only intervals that have lower boarders larger than
this value will be included in the resulting data frame.}
}
\value{
data.frame. A data frame containing the start and end times of the
intervals specified by the \code{brks} argument. Additionally the interval
length, interval mid-point and a factor variable of the intervals themselfs.
}
\description{
Given breaks, create start/end times and interval information
}

