% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lag-lead-helpers.R
\name{int_info2}
\alias{int_info2}
\title{Given interval break-points, provides data frame with information regarding
interval start-/end-points, lengths, midpoints,etc.}
\usage{
int_info2(x = c(0:12, seq(15, 55, by = 5), 61), min.time = 4)
}
\arguments{
\item{x}{A numeric vector of cut points in which the follow-up should be
partitioned in or object of class \code{ped}.}

\item{min.time}{Only intervals that have lower borders larger than
this value will be included in the resulting data frame.}
}
\description{
Given interval break-points, provides data frame with information regarding
interval start-/end-points, lengths, midpoints,etc.
}
