% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct-newdata.R
\name{make_newdata}
\alias{make_newdata}
\title{Construct a data frame suitable for prediction}
\usage{
make_newdata(ped, ..., expand = NULL, length.out = 25L)
}
\arguments{
\item{ped}{An object of class \code{ped} as returned by \code{\link[pam]{split_data}}.}

\item{...}{Further specifications of variables that should be set 
to a specific value.}

\item{expand}{A charachter vector of column names in \code{ped}.}

\item{length.out}{desired length of the sequence.  A
    non-negative number, which for \code{seq} and \code{seq.int} will be
    rounded up if fractional.}
}
\description{
Given a data set of class \code{ped}, returns a data frame that can be used 
as \code{newdata} argument in a call to \code{predict} and similar functions. 
When \code{expand==NULL} falls back to \code{\link{ped_info}}.
}
\details{
Details 
Extracts information from \code{ped}, using \code{\link{sample_info}}. 
If variables are specified with specific values in \code{...}, the values
in from \code{sample_info} will be overwritten. If variables provided 
in expand, these will be exanded from \code{min} to \code{max} using 
in \code{length.out} equidistant steps.
}
\examples{
\dontrun{
library(dplyr)
iris \%>\% make_newdata()
iris \%>\% make_newdata(Sepal.Length=5)
iris \%>\% make_newdata(Sepal.Length=c(5, 10), Sepal.Width=c(5, 5.1, 5.2))
iris \%>\% make_newdata(Sepal.Length=c(5, 10), expand="Sepal.Width", length.out=5)
iris \%>\% group_by(Species) \%>\% 
  make_newdata(Sepal.Length=c(5, 10), expand="Sepal.Width", length.out=5) \%>\% 
  print(n=30)
}
}
