% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-functions.R
\name{make_X}
\alias{make_X}
\title{Create model matrix from model object}
\usage{
make_X(object, ...)
}
\arguments{
\item{object}{a model object}
}
\description{
Create model matrix from model object
}
\keyword{internal}
