% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pam-package.r
\docType{package}
\name{pam}
\alias{pam}
\alias{pam-package}
\title{pam: Piece-wise exponential Additive Models}
\description{
pam provides functions and utilities that facilitate fitting 
Piece-wise Exponential (Additive) Models, including data transformation and 
other convenience functions for pre- and postprocessing as well as plotting.
}
\details{
The best way to get an overview of the functionality provided and how to 
fit PAMs is to view the vignettes included in this package. A summary 
of the vignettes content is given below:

\itemize{
\item \code{basics}: Introduction to PAMs and basic modeling.
\item \code{baseline}: Shows how to estimate and visualize baseline model 
(without covariates) and comparison to respective Cox-PH model.}
}
