% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervalInformation.R
\name{ped_info}
\alias{ped_info}
\title{Extract inetrval information and median/modus values vor covariates}
\usage{
ped_info(ped)
}
\arguments{
\item{ped}{An object of class \code{ped} as returned by \code{\link[pam]{split_data}}.}
}
\description{
Given an object of class \code{ped}, returns data frame with interval information, 
median values for numerical variables and modi for non-numeric variables 
in the data set.
}
\examples{
data("leuk2", package="bpcp")
leuk.ped <- split_data(Surv(time, status)~., data=leuk2, id="id")
ped_info(leuk.ped)
}
\seealso{
\code{\link[pam]{int_info}}, \code{\link[pam]{sample_info}}
}
