% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pec-helpers.R
\name{predictSurvProb.pamm}
\alias{predictSurvProb.pamm}
\title{Predict survival probabilities at specified time points}
\usage{
predictSurvProb.pamm(object, newdata, times, ...)
}
\arguments{
\item{object}{A PAMM model object.}

\item{newdata}{Data containing covariate information only, for which to
predict survival probabilities.}

\item{times}{Time points for which survival probabilities should be predicted.}

\item{...}{Arguments passed to \code{\link[pamm]{int_info}}.}
}
\description{
This function is needed to work with the \code{pec} package for estimation
of the time-dependent C-Index and Brier-Score (prediction error)
}
\keyword{internal}
