% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predictSurvProb.pamm}
\alias{predictSurvProb.pamm}
\title{S3 method for pamm objects for compatibility with package pec}
\usage{
\method{predictSurvProb}{pamm}(object, newdata, times, ...)
}
\arguments{
\item{object}{A fitted model from which to extract predicted survival
probabilities}

\item{newdata}{A data frame containing predictor variable combinations for
which to compute predicted survival probabilities.}

\item{times}{A vector of times in the range of the response variable, e.g.
times when the response is a survival object, at which to return the
survival probabilities.}

\item{...}{Additional arguments that are passed on to the current method.}
}
\description{
S3 method for pamm objects for compatibility with package pec
}
