% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpexp.R
\name{rpexp}
\alias{rpexp}
\title{Draw random numbers from piece-wise exponential distribution.}
\usage{
rpexp(n = 1, rate = 1, t = 0)
}
\arguments{
\item{n}{number of observations. If \code{length(n) > 1}, the length is
taken to be the number required.}

\item{rate}{vector of rates.}

\item{t}{vector of the same length as \code{rate}, giving the times at which
the rate changes. The values of \code{t} should be in increasing order.}
}
\description{
This is a copy of the same function from \code{rpexp} from package
\pkg{msm}.
Copied here to reduce dependencies.
}
\keyword{internal}
