% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct-newdata.R
\name{sample_info}
\alias{sample_info}
\alias{sample_info.data.frame}
\alias{sample_info.ped}
\title{Extract information of the sample contained in a data set}
\usage{
sample_info(x, ...)

\method{sample_info}{data.frame}(x, ...)

\method{sample_info}{ped}(x, ...)
}
\arguments{
\item{x}{A data frame (or object that inherits from \code{data.frame}).}

\item{...}{Further arguments passed to specialized methods.}
}
\description{
Given a data set and grouping variables, this function returns median values 
for numeric variables and modus for characters and factors.
}
\seealso{
\code{\link[pam]{split_data}}
}
