% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidiers.R
\name{tidy_fixed.gam}
\alias{tidy_fixed.gam}
\alias{tidy_fixed.coxph}
\alias{tidy_smooth}
\alias{tidy_smooth2d}
\title{Extract 1d smooth objects in tidy data format.}
\usage{
\method{tidy_fixed}{gam}(x, intercept = FALSE, ...)

\method{tidy_fixed}{coxph}(x, ...)

tidy_smooth(x, keep = c("x", "fit", "se", "xlab", "ylab"), ci = TRUE, ...)

tidy_smooth2d(x, keep = c("x", "y", "fit", "se", "xlab", "ylab", "main"),
  ci = FALSE, ...)
}
\arguments{
\item{x}{A model object.}

\item{intercept}{Should intercept also be returned?}

\item{...}{Currently not used.}

\item{keep}{A vector of variables to keep.}

\item{ci}{A logical value indicating whether confidence intervals should be
calculated and returned. Defaults to \code{TRUE}.}
}
\description{
Extract 1d smooth objects in tidy data format.

Extract 2d smooth objects in tidy format.
}
