% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidiers.R
\name{tidy_re}
\alias{tidy_re}
\title{Extract random effects in tidy data format.}
\usage{
tidy_re(x, keep = c("fit", "main", "xlab", "ylab"), ...)
}
\arguments{
\item{x}{ a fitted \code{gam} object as produced by \code{gam()}.}

\item{keep}{A vector of variables to keep.}

\item{...}{Further arguments passed to \code{\link[mgcv]{plot.gam}}}
}
\description{
Extract random effects in tidy data format.
}
\seealso{
\code{\link[stats]{qqline}}
}
