% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidiers.R
\name{tidy_smooth}
\alias{tidy_smooth}
\alias{tidy_re}
\title{Extract 1d smooth objects in tidy data format.}
\usage{
tidy_smooth(x, keep = c("x", "fit", "se", "xlab", "ylab"), ci = TRUE, ...)

tidy_re(x, keep = c("fit", "main", "xlab", "ylab"), ...)
}
\arguments{
\item{x}{ a fitted \code{gam} object as produced by \code{gam()}.}

\item{keep}{A vector of variables to keep.}

\item{ci}{A logical value indicating whether confidence intervals should be
calculated and returned. Defaults to \code{TRUE}.}

\item{...}{Further arguments passed to \code{\link[mgcv]{plot.gam}}}
}
\description{
Extract 1d smooth objects in tidy data format.

Extract random effects objects in tidy data format.
}
\seealso{
\code{\link[stats]{qqline}}
}
