% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidiers.R
\name{tidy_smooth2d}
\alias{tidy_smooth2d}
\title{Extract 2d smooth objects in tidy format.}
\usage{
tidy_smooth2d(x, keep = c("x", "y", "fit", "se", "xlab", "ylab", "main"),
  ci = FALSE, ...)
}
\arguments{
\item{x}{ a fitted \code{gam} object as produced by \code{gam()}.}

\item{keep}{A vector of variables to keep.}

\item{ci}{A logical value indicating whether confidence intervals should be
calculated and returned. Defaults to \code{TRUE}.}

\item{...}{Further arguments passed to \code{\link[mgcv]{plot.gam}}}
}
\description{
Extract 2d smooth objects in tidy format.
}
