% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyverse-methods.R
\name{tidyr_verbs}
\alias{tidyr_verbs}
\alias{fill}
\alias{fill_}
\alias{fill.ped}
\alias{fill.nested_fdf}
\title{\code{tidyr} Verbs for \code{ped}-Objects}
\usage{
\method{fill}{ped}(data, ..., .direction = c("down", "up"),
  keep_attributes = TRUE)

\method{fill}{nested_fdf}(data, ..., .direction = c("down", "up"),
  keep_attributes = TRUE)
}
\arguments{
\item{data}{an  object of class \code{ped}, see \code{\link{as_ped}}.}

\item{...}{A selection of columns. If empty, all variables are
selected. You can supply bare variable names, select all
variables between x and z with \code{x:z}, exclude y with \code{-y}. For
more options, see the \code{\link[dplyr:select]{dplyr::select()}} documentation. See also
the section on selection rules below.}

\item{.direction}{Direction in which to fill missing values. Currently
either "down" (the default) or "up".}

\item{keep_attributes}{conserve attributes? defaults to \code{TRUE}.}
}
\value{
A modified \code{ped} object.
}
\description{
See \code{tidyr} documentation of the respective functions for
  description and examples.
}
\keyword{internal}
