% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyverse-ped-methods.R
\name{tidyr_verbs}
\alias{tidyr_verbs}
\alias{fill}
\alias{fill_}
\alias{fill.ped}
\alias{fill_.ped}
\title{\code{tidyr} Verbs for \code{ped}-Objects}
\usage{
fill.ped(data, ..., .direction = c("down", "up"), keep_attributes = TRUE)

\method{fill_}{ped}(data, fill_cols, .direction = c("down", "up"))
}
\arguments{
\item{data}{an  object of class \code{ped}, see \code{\link{split_data}}.}

\item{...}{Specification of columns to fill. Use bare variable names.
Select all variables between x and z with \code{x:z}, exclude y with
\code{-y}. For more options, see the \link[dplyr]{select} documentation.}

\item{.direction}{Direction in which to fill missing values. Currently
either "down" (the default) or "up".}

\item{keep_attributes}{conserve attributes? defaults to \code{TRUE}.}

\item{fill_cols}{Character vector of column names.}
}
\value{
A modified \code{ped} object.
}
\description{
See \code{tidyr} documentation of the respective functions for
  description and examples.
}
